muzleflesh /rateo di fuoco

+++ battle game/battlegame/Assets/script/Soldier/editor_soldier/Soldier_Data.cs
@@ -14,5 +14,5 @@ public class Soldier_Data : ScriptableObject
     public float AreaDiVista;
     public float Esperienza;
     public float TempoDiReazione;
-    public float Vita;
+    public int Vita;
 }
